+++ kernel/timer.h
@@ -2,5 +2,5 @@
 #include <stdint.h>
 
 void timer_init(uint32_t freq);
-uint64_t timer_uptime_ms(void);
+uint32_t timer_uptime_ms(void);
